Fixes of HW02 (optional) InMemoryMealRepositoryImpl 1. method "save" throws NullPointerException, fixed. 2. method "delete" allow to delete alien meal, fixed. 3. method "get" simplified. 4. method "getAll" simplified. InMemoryUserRepositoryImpl 1. method "getAll" sorting data in a wrong order, fixed. 2. method "getByMail" simplified. MealRestController 1. method "getAll" refactored and simplified. MealServlet 1. getting filtered data moved to the method "doPost"

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -7,7 +7,6 @@ import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -31,8 +30,9 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
         } else {
-            Meal existingMeal4Update = repository.get(meal.getId());
-            if (!existingMeal4Update.getUserId().equals(userId)) return null;
+            if (this.get(meal.getId(), userId) == null) {
+                return null;
+            }
         }
         meal.setUserId(userId);
         repository.put(meal.getId(), meal);
@@ -41,34 +41,36 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        Meal meal = repository.remove(id);
-        return !(meal == null || !meal.getUserId().equals(userId));
+        if (this.get(id, userId) == null) {
+            return false;
+        }
+        repository.remove(id);
+        return true;
     }
 
     @Override
     public Meal get(int id, int userId) {
         Meal meal = repository.get(id);
-        return (meal == null || !meal.getUserId().equals(userId)) ? null : meal;
+        return (!meal.getUserId().equals(userId)) ? null : meal;
     }
 
     @Override
     public Collection<Meal> getAll(int userId) {
-        return getAll(userId, LocalDate.MIN, LocalDate.MAX, LocalTime.MIN, LocalTime.MAX);
+        return getAll(userId, LocalDate.MIN, LocalDate.MAX);
     }
 
     @Override
-    public Collection<Meal> getAll(int userId, LocalDate startDate, LocalDate endDate, LocalTime startTime, LocalTime endTime) {
+    public Collection<Meal> getAll(int userId, LocalDate startDate, LocalDate endDate) {
         List<Meal> result = new ArrayList<>(
                 repository
                         .values()
                         .stream()
                         .filter(meal -> meal.getUserId()==userId)
-                        .filter(meal -> DateTimeUtil.isBetween(meal.getDateTime().toLocalTime(), startTime, endTime))
                         .filter(meal -> DateTimeUtil.isBetween(meal.getDateTime().toLocalDate(), startDate, endDate))
                         .collect(Collectors.toList())
         );
         Collections.sort(result, (o1, o2) -> o2.getDateTime().compareTo(o1.getDateTime()));
-        return Optional.of(result).orElseGet(Collections::emptyList);
+        return result;
     }
 }
